JMAQS #106 - Selenium - CleanUp - 0h 4m -  Moved and Refactored BrowserType.
Module [JMAQS]:
  * BrowserType.java
  * SeleniumConfig.java
  * WebDriverFactory.java
  * SeleniumConfigTest.java
  * WebDriverFactoryTest.java
