+++ Framework/jmaqs-selenium/src/test/java/com/magenic/jmaqs/selenium/WebDriverFactoryTest.java
@@ -4,6 +4,7 @@
 
 package com.magenic.jmaqs.selenium;
 
+import com.magenic.jmaqs.selenium.constants.BrowserType;
 import com.magenic.jmaqs.selenium.constants.WebDriverFile;
 import com.magenic.jmaqs.utilities.helper.TestCategories;
 import java.util.HashMap;
@@ -73,7 +74,8 @@ public class WebDriverFactoryTest {
   public void getChromeDriverTest() throws Exception {
     ChromeDriver driver = null;
     try {
-      driver = (ChromeDriver) WebDriverFactory.getBrowserWithDefaultConfiguration(BrowserType.Chrome);
+      driver = (ChromeDriver) WebDriverFactory
+          .getBrowserWithDefaultConfiguration(BrowserType.CHROME);
       Assert.assertNotNull(driver);
     } finally {
       if (driver != null) {
@@ -86,7 +88,8 @@ public class WebDriverFactoryTest {
   public void getHeadlessChromeDriverTest() throws Exception {
     ChromeDriver driver = null;
     try {
-      driver = (ChromeDriver) WebDriverFactory.getBrowserWithDefaultConfiguration(BrowserType.HeadlessChrome);
+      driver = (ChromeDriver) WebDriverFactory
+          .getBrowserWithDefaultConfiguration(BrowserType.HEADLESS_CHROME);
       Assert.assertNotNull(driver);
     } finally {
       if (driver != null) {
@@ -99,7 +102,8 @@ public class WebDriverFactoryTest {
   public void getFirefoxDriverTest() throws Exception {
     FirefoxDriver driver = null;
     try {
-      driver = (FirefoxDriver) WebDriverFactory.getBrowserWithDefaultConfiguration(BrowserType.Firefox);
+      driver = (FirefoxDriver) WebDriverFactory
+          .getBrowserWithDefaultConfiguration(BrowserType.FIREFOX);
       Assert.assertNotNull(driver);
     } finally {
       if (driver != null) {
